# Phase 2-20260109T205517Z-3-001/Phase 2/Phase 2/swagger-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ComStag API",
    "description": "Organizations & Consumers platform",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://comstag-back.onrender.com",
      "description": "Generated server url"
    }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/success-story/{successStoryId}": {
      "put": {
        "tags": ["update-success-story-processor"],
        "summary": "Edit success story",
        "description": "1. Edit the success story of the current organization.\n2. If there are media to delete or to add upload them first.\n",
        "operationId": "editSuccessStory",
        "parameters": [
          {
            "name": "successStoryId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSuccessStoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      },
      "delete": {
        "tags": ["delete-success-story-processor"],
        "summary": "Delete success story",
        "operationId": "deleteSuccessStory",
        "parameters": [
          {
            "name": "successStoryId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/profile/reset-pass": {
      "put": {
        "tags": ["profile-reset-password-processor"],
        "summary": "Reset profile password",
        "description": "Reset password for the current user\n",
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/post/{postId}": {
      "get": {
        "tags": ["get-post-processor"],
        "summary": "get post details",
        "operationId": "getPost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      },
      "put": {
        "tags": ["update-post-processor"],
        "summary": "Update post",
        "description": "1. update post wth specific post id.\n2. Only active organization have access to this endpoint.\n3. upload media first if needed.\n",
        "operationId": "editPost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePostRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      },
      "delete": {
        "tags": ["delete-post-processor"],
        "summary": "delete post",
        "operationId": "deletePost",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/v1/success-story": {
      "post": {
        "tags": ["create-success-story-processor"],
        "summary": "Create success story",
        "description": "1. Create a success story for the current organization.\n2. Upload the media first.\n",
        "operationId": "createSuccessStory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SuccessStoryRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/post": {
      "post": {
        "tags": ["create-post-processor"],
        "summary": "Create post",
        "description": "1. Create a post for the current organization.\n2. Upload the media first.\n",
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
    "/v1/auth/reset-pass": {
      "post": {
        "tags": ["auth-reset-password-processor"],
        "operationId": "resetPassword_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/post/my-list": {
      "get": {
        "tags": ["list-posts-processor"],
        "summary": "list my posts",
        "description": "1. List my posts using access token only\n2. Only active organization have access to this endpoint\n",
        "operationId": "listMyPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "maximum": 100,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/post/list": {
      "get": {
        "tags": ["list-posts-processor"],
        "summary": "list posts",
        "description": "List posts of specific org and used when you view organization profile.\n",
        "operationId": "listOrgPosts",
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "maximum": 100,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/v1/post/latest": {
      "get": {
        "tags": ["list-posts-processor"],
        "summary": "list latest posts",
        "description": "List latest posts and used in home page\n",
        "operationId": "listLatest",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "maximum": 100,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateSuccessStoryRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "body": { "type": "string" },
          "deletedMediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          },
          "newMediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          },
          "deletedHashtagIds": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" },
            "uniqueItems": true
          },
          "newHashtagIds": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" },
            "uniqueItems": true
          }
        }
      },
      "ProfileResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "newPassword": { "type": "string", "minLength": 1 },
          "oldPassword": { "type": "string", "minLength": 1 }
        },
        "required": ["newPassword", "oldPassword"]
      },
...
      "UpdatePostRequest": {
        "type": "object",
        "properties": {
          "body": { "type": "string" },
          "deletedMediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          },
          "newMediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          }
        }
      },
...
      "SuccessStoryRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "maxLength": 255, "minLength": 0 },
          "body": { "type": "string", "minLength": 1 },
          "mediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "maxItems": 20,
            "minItems": 0,
            "uniqueItems": true
          },
          "hashtags": {
            "type": "array",
            "items": { "type": "integer", "format": "int64" },
            "maxItems": 10,
            "minItems": 1,
            "uniqueItems": true
          }
        },
        "required": ["body", "title"]
      },
      "PostRequest": {
        "type": "object",
        "properties": {
          "body": { "type": "string", "minLength": 1 },
          "mediaIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" },
            "uniqueItems": true
          }
        },
        "required": ["body"]
      },
...
      "AuthResetPasswordRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "newPassword": { "type": "string", "minLength": 1 },
          "verificationCode": { "type": "string", "minLength": 1 }
        },
        "required": ["newPassword", "verificationCode"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Paste your IdToken here",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}